Update patch set 2
Patch Set 2: Code-Review-1

(2 comments)

> Patch Set 2:
> 
> (1 comment)
> 
> I converted googletest to a git submodule.  Just so you don't think I'm crazy, there's a decent amount of "subtree > submodule" sentiment online similar to [1].  I also added a third_party/README.md and added some brief additional submodule steps to //README.md.
> 
> [1] https://codingkilledthecat.wordpress.com/2012/04/28/why-your-company-shouldnt-use-git-submodules/

OK, I can see there are some cons to submodules as well.  It looks like you can get into trouble if there is a version bump as well as local changes to the submodule, you have to be extra careful during the merge/rebase workflow.

The common case (at least for Chromium projects) is to commit to the upstream respository directly (not via a submodule) then roll the submodule downstream in a completely separate patch, so hopefully this won't be too much of a problem in practice.  If this turns out to be a hassle we can try subtrees or repo (maybe asking around to see if there is other advice too).

Patch-set: 2
Reviewer: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1