[WIP] Add ScreenListener embedder API
This currently includes a screen list and state machine implementation
but no mDNS code.  The intention is for mDNS to be a sibling of both
ScreenListener and ScreenPublisher, rather than owned by either, since
one instance is effectively shared between them.  This means that
ScreenListenerImpl (and eventually ScreenPublisherImpl) still needs a
driver which glues it to an mDNS implementation.

The likely-contentious part of this change is that the state machine is
asynchronous.  For example, calling Start() actually just queues a start
transition and returns, the intention being that the aforementioned
driver will both coalesce listener and publisher state machines and
perform resultant transitions in the task queue/event loop.

The alternative would be to make this synchronous, which would at least
involve giving the ScreenListenerImpl a pointer to its driver so it can
immediately request state changes.

The advantage of the asynchronous approach is that it might reduce
redundancy in steps the driver takes (this does not mean
redundancy/complexity in the code) to merge state machine transitions.
The disadvantage is that it presents a potential for getting too "far
ahead" of the real state.

Change-Id: I18699b871d13cb8fffbfb15ed4a1a0c9f20a7a5b