Update patch set 1
Patch Set 1:

(36 comments)

> Patch Set 1:
> 
> The main obstacle I see for making these asynchronous (e.g. void Send(..., void (cb*)(int result))) is platforms other than chromium where these are synchronous may end up with heavy recursion.  For example, setting up a socket with an asynchronous API but a minimal synchronous implementation might result in the following stack:
>     SetupSocket()
>     platform::CreateUdpSocket()
>     OnSocketCreated()
>     platform::BindUdpSocket()
>     OnSocketBound()
>     platform::SetUdpMulticastProperties()
>     OnMulticastPropertiesSet()
>     platform::JoinUdpMulticastGroup()
>     OnGroupJoined()
> 
> This might not be too bad, but we may want to add a recommendation that the platform should have its own event loop to avoid this.
> 
> What do you think about this option, with or without the event loop/synchronous recommendation?

Chromium's implementation uses a state machine to avoid recursion, i.e the callbacks update a state variable and then re-enter an event loop that dispatches to the next operation.  It's designed to handle either synchronous or asynchronous completion of these operations.  See the cast_channel implementation for an example of this.

These are tricky to get right.  It would simplify the code somewhat to have an implementation that is always asynchronous, but I am not entirely sure if that is possible on Linux.  In theory, there are some POSIX APIs for asynchronous IO (aio_read, aio_write, etc.) but Linux implementations seem to not be trusted at this point.

Patch-set: 1
Reviewer: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>