[Open Screen] Adds the NetworkServiceManager and dependencies.
This is the top level of the embedder API that allows the embedder
to instantiate and inject network services into the library.

The API for the ScreenListener service is also included in this change.
Other APIs are stubbed out and will be added in future patches.

This also fixes a bug in the Ipv4/v6Address constructor and adds a
boolean operator, so that a null address can be tested.

Change-Id: I49417c86c13d8f4351ffffe3123529d8b399e7ff