Update patch set 2
Patch Set 2:

(6 comments)

Overall looks good, would like to see if we can simplify a bit.

First, let's fix the API design by having 1:1 mapping between all possible states and transitions and the State enum exposed through the API (so we don't need SearchingFromRunning and SearchingFromSuspended).

Secondly, I'm not sure it makes sense to have implementations for the methods that initiate state transitions, because we don't know the details of how the implementation will implement those transitions (whether sync or async).

I think the design is using a queue of future transitions to provide some flexibility in this, but the simpler approach would seem to require the actual implementation to perform the state update when it's performed the task requested through the API successfully.

It does make sense to have a utility method ValidateNextState(transition) that any implementation can use to determine if the current state allows the requested transition (before the implementation starts the task) and UpdateState(transition, new_state) that revalidates the transition, modifies the current state and notifies any observers.

I believe that will allow an implementation (subclass) to implement the API in a sync or async manner.  Does that make sense?

Patch-set: 2
Reviewer: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>