Update patch set 1
Patch Set 1:

(12 comments)

This looks good overall.  Most of my comments are about simplifying the encoding and parsing code to reduce the number of repeated code patterns.   I think a fair amount of time could be spent on polishing that approach, but a few high level things stood out:

- Having a macro to to handle error returns from the CBOR library.
- Extracting functions to consume an expected tag or map key.
- Extracting functions to consume common value types as needed, like arrays of strings.

It's always tempting to go the route of code generation, especially if we add schemas for messages.  As we add more messages I suspect we'll want that eventually.  Let's see how the hand written code looks with a bit of refactoring first.

I did have a couple of functional questions:

- Is CBOR neutral on character encoding?  If so, we'll need to check for valid UTF-8 for both encoding and parsing.  Or we could go the route of URL-encoding to ASCII, but that generally speaking should not be necessary for new protocols that are 8-bit clean.  I did have a question about handling of embedded nulls.
- How does the caller manage sizing of the encoding buffer?

We'll want to add a test suite to exercise all the parsing and encoding logic thoroughly.  It might be useful to set up a test with known-good and known-bad message files that we can just load and iterate over.

We'll also want to set up fuzzer tests (for the whole library, really, but message parsing/encoding should be a focus).

I can file issues for those tasks to work on later :)

Patch-set: 1
Reviewer: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>