Merge api/impl and api/public BUILD.gn files
Several implementations of the classes in api/public need to access code
from api/impl (as you might expect).  In order to prevent a circular
dependency, these either need to use the Pimpl idiom (undesirable) or
both directories need to simply be built together.  This change
implements the latter option.

Change-Id: I35bc471b256e2986d42d472b6d8561960a3b0f3f