Add Mac toolchain, refactor GN build files, misc clean-ups.
1. Refactors the BUILDCONFIG-related files to allow for building on both
Linux and Mac (and leaves open the possibility for other platforms in
the future). Much of the new LOC are borrowed from the main Chromium
project. Also, ran `gn format` on all the touched BUILD.gn files.

2. Added Mac LLVM toolchain definitions. With this, everything compiles
on Mac, but executables will not successfully link due to missing Mac-
specific impl. Upcoming changes will address this.

3. Increased strictness of compiler flags to match Chromium C++
standards. This includes: Disabling C++ exceptions and RTTI, disallowing
exit-time destructors, and enabling all warning flags except for a few.
For third-party code, some warnings have been disabled to allow them to
build. Some of the increased strictness is only enforced when building
with LLVM, not GCC.

4. While working on third_party/chromium_quic/build/base/BUILD.gn, it
was noticed that the allocator shim defined there was enabled. Since a
well-behaved library should not attempt to override malloc(), I've
disabled that.

5. Added miu@chromium.org to AUTHORS. Hi! :)

Change-Id: I1a3dd43bfc34dea5c3d62f2d7745b80a8020612f