Fix CDDL optional member handling
This change adds better handling of optional members in CDDL structures
by allowing a single block of members to be optional (but in a
restricted way).  In the current message spec, this covers the
Presentation API messages.  To cover the remaining message spec, only
map support (i.e. not arrays) needs to be further improved.  Map support
will be improved as part of a future change that addresses out-of-order
keys.

Change-Id: I4242596c2734f5ec3a1bd1d151205d2017926269