Add mDNS API for updating TXT data
This change adds an API to MdnsResponderAdapter to update the TXT data
of a registered service that is being published.  It also removes the
friendly name API that would have used this (as it's been removed from
the spec).  The future use of the TXT update API will be for updating
the service certificate fingerprint or, more likely, updating the
metadata timestamp.  This also adds config parameters for the publisher
hostname and service instance.