Add Receiver + Connection implementations
This change adds an initial presentation::Receiver implementation as
well as a presentation::Connection implementation.  Although the latter
slightly depends on both the Controller and Receiver, the few places
where it needs the Controller are commented out for now, and will be
implemented when the Controller lands.

Cherry-pick of original branch by @btolsch.

Change-Id: I460793de3d6dbf7e33d7831d46518d49a794d2f7