Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I'm curious--since it sounds like we plan on maintaining this Go implementation (at the very least for W3C spec implementation reasons), do we have any quality checks that we can put in place? E.g. adding PRESUBMIT.sh steps?
> 
> I might just be missing something, but I think now's the right time to make sure that this code looks good, especially since it's likely to be the code that developers consuming our API refer to, instead of the C++ library.

At a minimum, we should add a script to verify that the code compiles successfully.  I looked briefly for adding the go compiler as a GN toolchain and didn't see much, so it might be a one-off script that is run via PRESUBMIT like you suggested :-)

I don't know if a lot of testing is necessary unless someone decides to ship it as production code.  Maybe we can socialize a policy to update this code whenever a spec change lands in GitHub.

Patch-set: 1
Reviewer: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>