Consolidate platform/BUILD.gn files, and add 'build_with_chromium'.
Consolidate the several tiny platform BUILD.gn files into one BUILD.gn
file for everything. This "flattening" is justified because there isn't
so much going on here that necessitates a more-complex hierarchy.

Also, add the 'build_with_chromium' global build override and stub-out a
new platform/chromium directory for all the future Chromium-embedder
adapter code. This is a starting-point for being able to build Open
Screen as either: a) standalone; or b) to be embedded in Chromium. Note
that this 'build_with_chromium' approach is used by several third-party
projects used by Chromium.

Change-Id: Ib3e526d6c557ee8a297e166d4b1c442d1a0ea5a7