Update patch set 1
Patch Set 1:

(1 comment)

Thanks for doing this - fewer files are nice.

Until we get an unreasonable number of flags, it might be simpler to just add the new flag to the declare_args in BUILDCONFIG.gn or platform/BUILD.gn.

It would be nice if we got the Chromium build rules that allow you to put the platform name in a .cc to get platform-specific source lists.  BUILD rule optimization is never done ;-)

Patch-set: 1
Reviewer: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>