Enable building demo executable on Mac OS X
Currently, building the demo is disabled on Mac OS X since there is
a duplicate definition bug between mDNSResponder and chromium_quic. This
patch removes the current fix, which is passing a special flag and
disabling builds on Mac, with a cross platform solution of renaming the
symbols in mDNSResponder's BUILD.gn as they are built. The demo build is
also reenabled on Mac, although follow up work is required to get the
demo working end to end on Mac, esp. around interfacing with Bonjou

Change-Id: I1480ec83ff5d73f0421b2b367e1b54d2d1821e3e