Remove verbose logging levels from OSP
Currently, we have this concept of numerical logging levels, however
since verbose logging is done relatively rarely, the levels are poorly
defined and unnecessarily complex. This patch removes the numeric
logging levels in favor of a single OSP_LOG_VERBOSE macro.

This patch also changes the demo to include a usage string, as well as a
-v flag that toggles between info and verbose logging levels.

Change-Id: I66850bd1d2bb65371547fad2880a18436919ac09