Update patch set 1
Patch Set 1:

(4 comments)

> Patch Set 1:
> 
> (7 comments)
> 
> > Patch Set 1:
> > 
> > (9 comments)
> > 
> > Here's some initial comments.  Also, you'll need to check in an updated cddl file.  Do you have one that you're testing with already?
> 
> I have an updated CDDL file locally, but I didn't want to swap it in yet for a few reasons:
> - there are the issues with the new spec I called out on the group chat
> - some constant names are re-used in that spec, which are causing compile-time issues
> - The new spec changes a LOT of property names, which result in a lot of compile-time failures outside of the tool
> 
> I am happy to swap in the new CDDL spec (was hoping to take that up once modifications to the tool are done), but it is a larger change that touches many other files and I suspect a couple changes will still need to be made to it, so I think it should be a separate CL
> 
> Thoughts?

The issues you mentioned with the spec will indeed have to be fixed, but in the meantime we need to make sure each commit compiles and runs correctly.  I was originally concerned because I was thinking this patch would enforce integer-keys-only.  Since that's not happening, it should just be compatible with the existing CDDL.  So no changes there are necessary in this CL.

It might be nice to add one integer key example with a test that ensures it actually serializes as an integer key, but I'd say that's optional for now.

Patch-set: 1
Reviewer: Gerrit User 1215804 <1215804@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>