Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (1 comment)
> 
> > Patch Set 3: Code-Review+1
> > 
> > (6 comments)
> > 
> > Change looks good but it raises some spec questions (issue link in the comments)
> 
> More of a general question, but is there a reason we are relying on maintaining a counter manually for request-ids instead of generating a new random uint64 for each request? Is the guaranteed ordering also being used for ordering the media data sent over-the-wire?

QUIC guarantees in-order delivery of messages and in general we need to process control messages in-order to implement the protocol correctly.  For example, closing a presentation connection should prevent further handling of application messages on the connection.

We could generate random request IDs for matching requests and responses, and have a separate counter to maintain message order internally, but it seems simpler to use a monotonically increasing ID for both.

It also eliminates the ordering dependency on a *specific* QUIC connection, which allows endpoints to support connection migration while processing messages in-order.

For media data, the spec we are proposing will tag video frames with a frame id that in most cases will also be the transmission order, but things get more complicated because of inter-frame dependencies, so out-of-order transmission could also happen. [1]  

[1] https://docs.google.com/presentation/d/1-jAG9ixuSBMqSyUZVQ_9GdIPQB8fmGOpqyyQUVgzHPs/edit#slide=id.g466a43ee95_0_1388

I think, we do need to figure out the right way to initialize the per-endpoint counter for control messages, which we can sort out in the spec issue.

Patch-set: 3
Reviewer: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>