Track request IDs by endpoint ID
This change adds a client/server-scoped tracker for request IDs that is
indexed by endpoint IDs.  This is necessary because any API using CBOR
messages shares the |request_id| field which must be unique within a
connection.

NOTE: This was originally uploaded by btolsch@ as
https://chromium-review.googlesource.com/c/openscreen/+/1495225.  I'm
submitting it on his behalf with a few TODOs added while he is OOO.

Change-Id: I2828ef9dc9c94eebe1fc1ada28df2b36f8ff5054