Update patch set 3
Patch Set 3:

(11 comments)

I've had to fight with logging libraries a lot over the last few years, so I've developed a pretty strong opinion that we should have a wrapper around the outside calls (be that direct printf statements, calls to //platform/api/logging.h methods, or anything else) that takes in the most general form we can support. If you're really against this wrapper idea, please grab me and we can discuss it in person, but beyond having a wrapper I am open to changing pretty much anything

Mark, to your idea to use DVLOG(), is this preferred to using the OSP_LOG_... methods and writing to the provided stream directly (approach I have here)? Since this program is used at build-time, I prefer not to have a build flag for output since intermittent failures may not repro every time. It seems safer to have the output directed to a file for every build (or redirecting stdout as mentioned in this CL), so any build failure can be debugged. Thoughts?

Patch-set: 3
Reviewer: Gerrit User 1330244 <1330244@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>