Update patch set 2
Patch Set 2:

(7 comments)

Overall looks good.  All of the declarations make sense, but I did not manually verify that each message matched the spec though.  

I did have some questions about the tagging.   In the spec group, we debated whether to represent each message with a registered CBOR tag, or to use a single CBOR tag and type-prefix the CBOR messages.  From the meeting minutes [1], it looks like we are awaiting implementation feedback as one input into the decision, so this might be as good a time as any to hash it out.

[1] https://www.w3.org/2018/10/25-webscreens-minutes.html#x20

Regardless, tags we use in our CDDL shouldn't conflict with globally defined tags for CBOR or in IANA.  If we're using tags in our current implementation, we should pick an unreserved range in the 1000's and then add the type number from the spec.

Patch-set: 2
Reviewer: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>