Add start/terminate presentation support
This change adds a Controller implementation of StartPresentation and
OnPresentationTerminated, as well as other dependent changes.  This
includes some Connection changes and small Receiver fixes.  This change
also updates the demo to allow starting and terminating a presentation.

NOTE: This was originally uploaded as
https: //chromium-review.googlesource.com/c/openscreen/+/1496795
Change-Id: I078ec0672a192cfb5ad36eb6b9639077f08770c4