Update patch set 19
Patch Set 19:

(11 comments)

Sorry for not getting to this before it was merged.  Most of my comments are readability improvements, but it seems like there is one bug (the last comment) and there's one thing we could do to improve the use of this in unit tests that test things that use TaskRunners.

Patch-set: 19