Update patch set 1
Patch Set 1:

(4 comments)

On the topic of the fake clock:

I think we need to solve the issue of making it possible to run test of things that use the TaskRunner, but to do so with a fake clock.  But RunUntilStopped does not work with a fake clock because run_loop_wakeup_.wait_until does not work with a fake clock.

I'm not sure the best way to solve this, but it's going to be an issue that we'll have to solve eventually.  Perhaps the right thing to do for now is to just put a TODO in the file and remember to come back to it later.

Patch-set: 1
CC: Gerrit User 1125619 <1125619@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>