Always use clock now function ptr in TaskRunner
Currently, the TaskRunner relies on the STL's understanding of the
system clock for some things, like the wait predicate, and uses the
actual platform::Clock for some unit tests. This causes unpredictable
behavior in our build bots, as well as violating best practices.

This patch moves to using a platform::ClockNowFunctionPtr in the
TaskRunnerFactory, always calling that FunctionPtr to get time updates
in the TaskRunnerImpl, and always passing a fake clock in unit tests.

Change-Id: I054f176ff34ee1bbd841e0f6cffef01de6af1af8