Update patch set 2
Patch Set 2:

(4 comments)

Can you explain the purpose of the "fake member"?  Does the data model of the AST not have a value for an empty group?  It seems like it would be easier to update the AST with a new node type than manage the insertion and removal of placeholder members.

Patch-set: 2
Reviewer: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>