Fix demo bugs + add demo documentation
This change fixes bugs in every command of the demo and adds a README.md
specifically for the demo.  A few unit tests were added to cover bugs
found while fixing the demo as well.  This should put us in a good
position to make the demo the default binary built by new developers
instead of hello_world.

Bug: openscreen:34
Change-Id: Ib76e200870197a37b2a7d6ca0a967502123a9cf6