Cast Streaming: RTP wire-formatting/parsing.
Adds RTP packet serialization for Cast Streaming. These are the packets
that carry the payload data and related timing/sequence information for
a media stream.

Most of this change is about testing:

1. Unit testing of packet parsing using hard-coded binary input.

2. Testing of "packetizing" a frame into one or more RTP packets, which
are then re-parsed and checked for original data values.

3. Fuzz testing of the packet parser, to prevent crashing, memory leaks,
etc., in the face of both expected and unexpected/corrupt inputs.

Change-Id: I449274bade27900dc06fe879f8645d8a551426a4