Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Not sure if you wanted to land this before or after the move to osp/.  Let me know so we can coordinate.

Probably before, assuming that the name is good. I can't think of a better name than osp_base, since we are keeping the entire repository as openscreen, osp makes sense even if it is shared IMO.

If we land this before the move, it will be an easy move. After will be a pain.

Patch-set: 3
Reviewer: Gerrit User 1323871 <1323871@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>