Cast Streaming RTCP Part 1: Sender Reports
Adds common header and report block RTCP packet parsing and
serialization, and uses this to implement the parsing and serialization
of RTCP Sender Reports. This patch includes all of the sender-to-
receiver RTCP messaging. (The upcoming Part 2 patch will build upon this
work and add all the receiver-to-sender RTCP messaging.)

Other notables:

1. Updated InspectPacketForRouting() for quick identification of RTCP
   packets.
2. Unit- and Fuzz-testing of all new parsing code.
3. Resolved the NtpTimestamp 2036 problem (it's now a 2106 problem);
   because a "warning DCHECK" was triggered by the fuzz testing.

Change-Id: Ibb87b59fbe7d6ce5661f16dcff46ac29fecef9d7