Clean up logging dependencies
Currently, our CDDL executable requires logging definitions that must be
provided by the embedder. Since CDDL is an executable target, it is
linked separately, before the embedder's logging implementation can be
build and linked. On top of this, we probably don't even want it using
the embedder's logging system... This patch changes the CDDL executable's
behavior to instead depend on the default logger, even when built inside
of an embedder.

Change-Id: I6a55e6d3779f2fdff0677972c48744d7f20f8918