Add TaskWaiter interface to TaskRunnerImpl
This change adds the option to use a TaskWaiter in TaskRunnerImpl
instead of a condition variable.  This allows for a more  cooperative
wait than a condition variable in the case where you may only want to
wait on something like a select/poll/epoll call.  This is not currently
going to be used but is part of one of our design alternatives for
integrating network polling into TaskRunner.

Bug: openscreen:46
Change-Id: I041117588c293230a138b45bb846db8526104c32