Update patch set 3
Patch Set 3:

(3 comments)

Adding some details about how I plan to use these changes, since I'm realizing that probably isn't as clear as I'd intended

My plan is to have an object which implements NetworkRunner interface (Brandon's CL), and have that own an instance of TaskRunner and of my new NetworkLoop class (which will be the only user of EventWaiter, to abstract away interacting with sockets). Then it can just forward all calls on NetworkLoop methods to NetworkLoop, Task executiom methods to TaskRunner, and all it's responsible for is running the EventWaiter Wait() loop on thread 2 (timeout TBD, but needs one to ensure new callbacks are picked up), and ensuring that the NetworkLoop knows about TaskRunner running on thread 1 (so NetworkLoop runs all of its callbacks on TaskRunner's thread 1).

It's fine if this in-between-state CL never gets checked in and we instead wait for the full async implementation - I just thought it was getting a bit large to proceed without one

Patch-set: 3
Reviewer: Gerrit User 1330244 <1330244@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0