Fix 32-bit build
Currently, due to an incorrect sizeof operation, the cddl parser
works properly on 64-bit (due to sizeof a char pointer being 8 bytes,
which is coincidentally the exact length of a particular string),
but doesn't work on 32-bit, since the char pointer is instead 4 bytes.

This patch resolves this issue by changing the type and using strlen
instead of sizeof. Logging is slightly improved as well.

Change-Id: Ifaf5e75ead5fd0325bb411689d8bf1d5669d331a