Update patch set 3
Patch Set 3:

(21 comments)

It sounds like it's not clear if this patch is moving forward in its current form.  Regardless I went ahead and reviewed it.  I don't have the context to know if we should remove EventWaiter in one fell swoop, or have an interim step where we move it out of platform/api and make it an implementation detail of the platform layer.

I am concerned with the current approach to the network loop that keeps callbacks and buffers outside of the socket objects - it results in some inefficient and risky code.  I would like to explore options that allow this state to be bound inside the socket objects - perhaps by moving the platform-specific pieces out into a separate header.

If we agree that's the right approach it would make sense to do in a separate change.

Patch-set: 3
Reviewer: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>