Update patch set 12
Patch Set 12:

(7 comments)

The label compression/decompression code is complicated and looks like it could be a potential exploit vector (it's basically pointer serialization).

I have not observed any compression in use by Cast devices in the field.  Since it's not required by the mDNS spec, what would the impact be of just not implementing it?

If we leave it in, it should fuzz tested to shake out any issues.

Patch-set: 12
Reviewer: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>