Update patch set 4
Patch Set 4:

(24 comments)

All I really focused on for this pass was the documentation :-)
Which raised a couple of questions we might want to chat about F2F.
Once that is sorted I will start to look through the other pieces.

PS.
This appears to be relevant documentation for Gitiles markdown, to help with some of the formatting stuff.


https://github.com/google/gitiles/blob/master/Documentation/markdown.md

Patch-set: 4
Reviewer: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>