Fixup CDDL builder python script
Currently, the script has a bug where if you are on a machine without
clang format installed, and not in verbose mode, it will fail to get the
path to clang format and then try to use it anyway, causing a crash.
This patch fixes this behavior, as well as a little bit of cleanup in
this script.