Update patch set 4
Patch Set 4:

(5 comments)

I started to review the code and realized, this isn't the design we agreed on at the team meeting a few weeks back. tl;dr: The networking layer needs to be separate from the TaskRunner implementation, and not the implementation of it itself.

FWIW, here are the comments I had so far:

Patch-set: 4
CC: Gerrit User 1002076 <1002076@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>