Update patch set 2
Patch Set 2:

(30 comments)

Thanks for breaking this up into smaller changes.  I did not make it to the platform/posix stuff yet.

I think I still need a picture of the threading model here.  It looks like the task runner is for the main OSP thread and the network loop is run by an I/O thread.  And the OSP thread needs to register sockets/callbacks with the network loop.  However the network loop design does not look fully re-entrant - are some methods intended to be called by only the I/O thread?

I don't like the design that loops over all the data structures in the network loop each time there is a socket event.  This kind of ignores the entire point of select() which is to tell you which sockets specifically need attention and seems like a perf issue.

Moreover holding/releasing the lock in the middle of Wait() is problematic because there's no guarantee that the set of sockets you're waiting on is the same that you have callbacks for.

Alternative:

1. At the beginning of Wait(), make a copy of the socket and callback pointers while holding the mutex.  Release mutex after copy.
2. Pass sockets into the event waiter
3. Event waiter returns sockets with events
4. Network loop invokes callbacks on those sockets

The caller (OSP thread) would need to understand that adding or removing sockets from the loop will not take effect until the next invocation of Wait().

Does that make sense?

Patch-set: 2
Reviewer: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>