Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (30 comments)
> 
> Thanks for breaking this up into smaller changes.  I did not make it to the platform/posix stuff yet.
> 
> I think I still need a picture of the threading model here.  It looks like the task runner is for the main OSP thread and the network loop is run by an I/O thread.  And the OSP thread needs to register sockets/callbacks with the network loop.  However the network loop design does not look fully re-entrant - are some methods intended to be called by only the I/O thread?
> 
> I don't like the design that loops over all the data structures in the network loop each time there is a socket event.  This kind of ignores the entire point of select() which is to tell you which sockets specifically need attention and seems like a perf issue.
> 
> Moreover holding/releasing the lock in the middle of Wait() is problematic because there's no guarantee that the set of sockets you're waiting on is the same that you have callbacks for.
> 
> Alternative:
> 
> 1. At the beginning of Wait(), make a copy of the socket and callback pointers while holding the mutex.  Release mutex after copy.
> 2. Pass sockets into the event waiter
> 3. Event waiter returns sockets with events
> 4. Network loop invokes callbacks on those sockets
> 
> The caller (OSP thread) would need to understand that adding or removing sockets from the loop will not take effect until the next invocation of Wait().
> 
> Does that make sense?

Let's discuss the threading model on Monday. I am under the impression that the NetworkLoop API should be fully thread-safe regardless of calling thread, so I may be missing something and would like to better understand what that is

Regarding looping over all data structures, I believe this to be a limitation of how the SELECT(...) API works. This is how we were performing the check in the previous EventWaiter code, and I do not see any alternative approach detailed in the FD_SET documentation: https://linux.die.net/man/3/fd_set
It would be possible to pass the sockets to the NetworkWaiter and get back the sockets that changes, but I could not come up with an implementation of that which was actually any more performant, given that we can't loop over an FD_SET

Regarding the NetworkLoop lock, I was under the impression that this had been discussed during the meeting between you, and Peter, and I, and that we had decided it was okay if the set of Sockets had changed or one of the Sockets had its callback replaced mid-wait-loop.
Let's discuss this further in person on Monday, as I don't see why having the callbacks or sockets change mid-loop actually causes any problems.

Patch-set: 2
Reviewer: Gerrit User 1330244 <1330244@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>