Fix noexcept ABI mangling issues and remove MAYBE_NOEXCEPT macro.
Removes the MAYBE_NOEXCEPT hack from the code base and instead adds
compiler flags to mute both GCC and Clang's warnings about ABI
incompatibilities when changing over to C++17. This is not relevant to
our project since the same compiler will always be used to build all C++
modules that are linked together.

Change-Id: Id7618b3931b4288f243799c4291e0dec71723228