Change chromium bots to trigger on chromium tree
Originally, we planned to have the chromium fyi bots only run for each
new openscreen commit, but against chromium HEAD.  The checked-in
gclient config for the chromium recipe isn't intended for that though,
as it was copied from v8's config.  This results in the bots not finding
a valid repo URL when they are triggered.

I believe the two ways to fix this are:
 1. Trigger on the chromium tree instead, so the bot is technically
    getting a chromium trigger, but then openscreen is set to HEAD by
    the current config in //s/s/r/chromium/gclient_config.py.
 2. Change the config in //s/s/r/chromium/gclient_config.py to rewrite
    the gclient solution for the current trigger.  It's unclear how to
    do this at the moment.

This change takes approach 1 since, in addition to being easier, it has
the added benefit of tracking our compile status in chromium without
openscreen even changing.

Bug: 951615
Change-Id: Ib34ffdf0ffb1f8abd5205b53348c93d8ba066194